refactor: rename consensus/light_client.rs -> consensus/processor.rs
Two files named light_client.rs at different layers — the public facade
(src/light_client.rs: LightClient/UpdateOutcome) and the internal engine
(src/consensus/light_client.rs: LightClientProcessor/LightClientStore) — was a
navigability trap flagged in the architecture map. Rename the engine to
processor.rs so the filename matches the type inside and the layer is obvious.

git mv + update the mod declaration, the two `crate::consensus::light_client`
imports, and the consensus/README unit-test table. Kept light_client_spec_tests.rs
as-is: that name points at the Ethereum spec's "light_client" test category
(matching the fixture dirs), not the impl file. Removed the now-resolved
navigability-wrinkle note from the architecture map.

Pure rename, no logic change; all tests pass, production clippy clean.
