+++ src/README.md
@@ -28,7 +28,7 @@ flowchart TD
 | `consensus/` | engine | SSZ / Merkle / BLS verification (private) |
 | `light_client` | facade | `LightClient`, `UpdateOutcome` (public entry) |
 
-**Facade vs engine.** `LightClient` (`src/light_client.rs`) is a thin public wrapper; the real work lives in `LightClientProcessor` (`src/consensus/light_client.rs`, `pub(crate)`). `process_update` delegates to the processor and wraps its `bool` into the richer `UpdateOutcome`. Consumers touch only the facade — `consensus/` is private. (Two files named `light_client.rs` at different layers is a known navigability wrinkle.)
+**Facade vs engine.** `LightClient` (`src/light_client.rs`) is a thin public wrapper; the real work lives in `LightClientProcessor` (`src/consensus/processor.rs`, `pub(crate)`). `process_update` delegates to the processor and wraps its `bool` into the richer `UpdateOutcome`. Consumers touch only the facade — `consensus/` is private.
 
 **The `types` umbrella spans two layers.** `types::primitives` sits *below* config (leaf aliases, no deps); `types::consensus` sits *above* it (its types carry a `&ChainSpec`). So `config` depends on `types::primitives` while `types::consensus` depends on `config`, which makes the crate-level `config ↔ types` edge *look* circular. It isn't — the real order is `primitives → config → consensus`; only the shared `types` name blurs it.
